latest as of Sep7 2018
Files are generated to track the source of issues with plastic strain. To avoid generate any file that seems to increase the computing time, please delete everything between "debug starts" and "debug ends" in the file "fl_rheol.f90". A cleaner version of FLAC will be pushed after this issue is fixed.